reduce js loaded on wiki documents
reduce from 705KB to 302KB minified but uncompressed
